api/requirements.txt
fastapi
uvicorn
joblib
pandas
scikit-learn
google-generativeai
python-dotenv
pydantic
